+++ src/index.js
@@ -31,6 +31,8 @@ import { Picker, CityPicker, PickerGroup } from './components/picker';
 import Badge from './components/badge';
 //non standard
 import { Popup, PopupHeader } from './components/popup';
+import PullToRefresh from './components/ptr';
+import InfiniteLoader from './components/infiniteloader';
 
 export default {
     version,
@@ -116,5 +118,7 @@ export default {
 
     //non-standard
     Popup,
-    PopupHeader
+    PopupHeader,
+    PullToRefresh,
+    InfiniteLoader
 };